src/fase3mp/FabricaCazador.java
package fase3mp;

public class FabricaCazador extends FabricaPersonajes{
    
    
    public Personaje crearPersonaje(String nombre,Habilidad habilidadPersonaje,Arma[] armas,Arma[] armasActivas,Armadura[] armaduras, 
    Armadura armaduraActiva,Esbirro[] esbirros,int Salud,int Poder,Debilidad[] debilidades,Fortaleza[] fortalezas){
        
        Cazador cazador = new Cazador(nombre, habilidadPersonaje, armas, armasActivas, armaduras, armaduraActiva, esbirros, Salud, Poder, debilidades, fortalezas);
        return cazador;
    }
    
}
